refactor(esbabbler): move CallParticipantTileProps interface to models
