+++ packages/app/app/composables/message/room/call/useCallParticipantTiles.ts
@@ -1,5 +1,5 @@
 import type { CallParticipant } from "#shared/models/room/call/CallParticipant";
-import type { CallParticipantTileProps } from "@/components/Message/Content/Call/Participant/Tile.vue";
+import type { CallParticipantTileProps } from "@/models/message/room/call/CallParticipantTileProps";
 
 import { authClient } from "@/services/auth/authClient";
 import { useCallStore } from "@/store/message/room/call";
